0.2

+++ app/src/main/java/com/example/cameraapp/list_of_photos.java
@@ -6,13 +6,19 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import android.content.Intent;
 import android.net.Uri;
+import android.os.Build;
 import android.os.Bundle;
 import android.widget.Button;
+import android.widget.Toast;
 
 import com.example.cameraapp.DB.Photos;
 import com.example.cameraapp.DB.PhotosDB;
 import com.example.cameraapp.customAdapter.CustomAdapter;
 
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -27,16 +33,12 @@ public class list_of_photos extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_list_of_photos);
 
+        init();
         initRecycleView();
         new_photo();
     }
 
-    private void initRecycleView() {
-        RecyclerView recyclerView = findViewById(R.id.recycle_view);
-        recyclerView.setLayoutManager(new LinearLayoutManager(this));
-        customAdapter = new CustomAdapter(this);
-        recyclerView.setAdapter(customAdapter);
-        customAdapter.setPhotosList(photosList);
+    protected void init() {
         back_button = findViewById(R.id.backButton);
         delete = findViewById(R.id.delete);
 
@@ -46,12 +48,29 @@ public class list_of_photos extends AppCompatActivity {
 
         delete.setOnClickListener(view -> {
             if (photosList.size() != 0) {
-                db.photosDAO().delete(photosList.get(0));
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+                    try {
+                        Toast.makeText(this, "File was also deleted", Toast.LENGTH_LONG).show();
+                        Files.deleteIfExists(
+                                Paths.get(photosList.get(0).path));
+                    } catch (IOException e) {
+                        throw new RuntimeException(e);
+                    }
+                    db.photosDAO().delete(photosList.get(0));
+                }
             }
 
             photosList = db.photosDAO().getAllPhotos();
-            customAdapter.setPhotosList(photosList);
+            customAdapter.deleteItem(0);
         });
+    }
+
+    private void initRecycleView() {
+        RecyclerView recyclerView = findViewById(R.id.recycle_view);
+        recyclerView.setLayoutManager(new LinearLayoutManager(this));
+        customAdapter = new CustomAdapter(this);
+        recyclerView.setAdapter(customAdapter);
+        customAdapter.setPhotosList(photosList);
     }
 
     private void new_photo() {
